ClickWeave-Py/requirements.txt
customtkinter>=5.2.0
pynput>=1.7.6
pyautogui>=0.9.54
keyboard>=0.13.5
pillow>=10.0.0
mss>=9.0.1
apscheduler>=3.10.4
pydantic>=2.5.0
psutil>=5.9.0
pytest>=7.4.0
pyinstaller>=6.3.0